requirements.txt
# banking-dispute-assistant-v1 Requirements

# Core dependencies
streamlit==1.28.1
openai==1.51.0
python-dotenv==1.0.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
faker==20.1.0

# Async processing
aiohttp==3.9.1

# Utilities
pydantic==2.5.2

# Development
pytest==7.4.3
black==23.12.1
flake8==6.1.0